server/requirements.txt
python-dotenv
openai
mysql.connector
flask
requests